+++ lib/src/core/state.dart
@@ -1,14 +1,17 @@
 import 'package:flutter/material.dart';
 
 class SidebarState extends InheritedWidget {
-  /// Wether the sidebar is collapse or not
+  /// Whether the sidebar is collapse or not
+  ///
+  // ignore: prefer_typing_uninitialized_variables
   final isCollapse;
 
-  const SidebarState({super.key, required Widget child, this.isCollapse = false})
+  const SidebarState({super.key, required Widget child, this.isCollapse})
       : super(child: child);
 
   @override
-  bool updateShouldNotify(SidebarState oldWidget) => isCollapse != oldWidget.isCollapse;
+  bool updateShouldNotify(SidebarState oldWidget) =>
+      isCollapse != oldWidget.isCollapse;
 
   static SidebarState? of(BuildContext context) =>
       context.dependOnInheritedWidgetOfExactType<SidebarState>();